Don't mask errors when downstream parsing fails
If the client of this library mistakenly gets a NPE when parsing a status
update the exception is hidden in the logging. That makes debugging near
impossible.

So this is a 3-part fix.

1) Log the actual exceptions instead of the localizedErrorMessage.
2) Split apart the message parsing from the message handling. If a parse
error happens we'll close the connection. If the handler fails, just ignore
it and report the error.
3) Catch all exceptions in both phases - parsing and error handling and act
accordingly based on the phase you're in.